Updated index.js to be more of a db file, separated out helper functions
